projects/subRecon/requirements.txt
aioquic==1.2.0
airdrop-ng==1.1
airgraph-ng==1.1
archinfo==9.2.134
asgiref==3.8.1
attrs==23.2.1.dev0
autocommand==2.2.2
Automat==22.10.0
bandit==1.8.3
bcrypt==4.2.1
bidict==0.23.1
bitarray==3.0.0
bitstring==4.3.0
blinker==1.8.2
bluepy==1.3.0
Brlapi==0.8.6
Brotli==1.1.0
btrfsutil==6.13
capstone==5.0.5
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
colorama==0.4.6
colored-traceback==0.3.0
constantly==23.10.4
contourpy==1.3.1
cramjam==2.9.1
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
EditorConfig==0.12.4
eventlet==0.38.0
filebytes==0.10.2
filelock==3.17.0
Flask==3.1.0
Flask-Compress==1.14
Flask-SocketIO==5.3.6
fonttools==4.56.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==7.2.1
incremental==22.10.0
intervaltree==3.1.0
itemadapter==0.8.0
itemloaders==1.3.2
itsdangerous==2.1.2
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.5
jmespath==1.0.1
jsbeautifier==1.15.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaitaistruct==0.10
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kiwisolver==1.4.5
ldap3==2.9
libtorrent==2.0.11
lit==19.1.7.dev0
louis==3.32.0
lxml==5.3.1
Mako==1.3.9.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat2==0.13.5
matplotlib==3.10.0
mdurl==0.1.2
meson==1.7.0
mitmproxy==11.1.0
mitmproxy_linux==0.11.4
mitmproxy_rs==0.11.4
more-itertools==10.5.0
msgpack==1.0.5
mutagen==1.47.0
netsnmp-python==1.0a1
numpy==2.2.3
onionshare==2.6.2
onionshare_cli==2.6.2
packaging==24.2
paramiko==3.5.0
parsel==1.8.1
passlib==1.7.4
pefile==2024.8.26
pillow==11.1.0
platformdirs==4.3.6
plumbum==1.9.0
ply==3.11
pooch==1.8.2
Protego==0.3.1
protobuf==5.29.2
psutil==7.0.0
pt==1.0.0
publicsuffix2==2.20191221
pwndbg==2025.2.19
pwntools==4.14.0
pwquality==1.4.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.27.0
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
PyDispatcher==2.0.7
pyelftools==0.31
pygccxml==2.6.1
Pygments==2.19.1
PyGObject==3.50.0
pylsqpack==0.3.18
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyperclip==1.9.0
pypng==0.20231004.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.8.2.1
pysmi==0.3.4
pysnmp-lextudio==6.1.2
PySocks==1.7.1
python-dateutil==2.9.0
python-engineio==4.9.1
python-ptrace==0.9.9
python-snappy==0.7.2
python-socketio==5.8.0
pyvex==9.2.123
pyxdg==0.28
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
queuelib==1.7.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
ROPGadget==7.6
ropper==1.13.10
routersploit==3.4.6
rpds-py==0.22.3
rpyc==6.0.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scapy==2.6.1
scipy==1.15.2
Scrapy==2.12.0
service-identity==24.2.0
setproctitle==1.3.4
setuptools==75.8.0
shiboken6==6.8.2.1
shiboken6-generator==6.8.2.1
shodan==1.31.0
six==1.17.0
smmap==5.0.1
sortedcontainers==2.4.0
ssh-audit==3.3.0
sshtunnel==0.4.0
sshuttle==1.1.2
stem==1.8.3
stevedore==5.4.0
tabulate==0.9.0
TBB==0.2
thrift==0.20.0
tldextract==5.1.3
tornado==6.4.2
tqdm==4.67.1
Twisted==24.3.0
typing_extensions==4.12.2
unicorn==2.1.2
Unidecode==1.3.8
urh==2.9.6
urllib3==2.3.0
urwid==2.6.15
volatility3==2.7.0
w3lib==2.1.2
waitress==3.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.0
wifite==2.7.0
wsproto==1.2.0
XlsxWriter==3.1.9
xortool==1.0.0
yara-python==4.5.1
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
